test %100 norm %100

+++ include/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mbugday <mbugday@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/21 14:37:39 by mbugday           #+#    #+#             */
-/*   Updated: 2022/03/21 18:29:50 by mbugday          ###   ########.fr       */
+/*   Updated: 2022/03/27 02:52:11 by mbugday          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int		ft_printf(const char *str, ...);
 int		ft_formats(va_list args, const char format);
 int		ft_print_char(int c);
 int		ft_print_str(char *str);
-int		ft_print_ptr(unsigned long long ptr);
+int		ft_print_ptr(unsigned long ptr);
 int		ft_print_nbr(int n);
 int		ft_print_unsigned(unsigned int n);
 int		ft_print_hex(unsigned int num, const char format);
